+++ src/ctreeskit/arraylake_tools/__init__.py
@@ -5,10 +5,6 @@ Arraylake Tools - Tools for migrating from Geotiffs -> Arraylake
 from .common import ArraylakeDatasetConfig
 from .ingest import AnnualRasterIngester
 
-__version__ = "0.1.0"
-__author__ = "Naomi Provost"
-__email__ = "nprovost@ctrees.org"
-
 __all__ = [
     "ArraylakeDatasetConfig",
     "AnnualRasterIngester",